Solutions to day 3's spiral challenge
This was not my favorite & I churned on it quite a bit. I also wrote a
ton of code that I trashed because I didn't need it. Ideally, I should
have started with the logic of how to move a point on the board without
really contemplating a way to avoid walking the spiral.